warehouse dev + mail service

+++ warehouse/src/main/kotlin/it/polito/wa2/group17/warehouse/dto/SellRequest.kt
@@ -5,6 +5,6 @@ import javax.validation.constraints.NotNull
 
 data class SellRequest(
     @field:NotNull val productID: Long,
-    @field:NotNull @field:Min(0) val quantity: Long,
+    @field:NotNull @field:Min(0) val quantity: Int,
     var warehouseID: Long?
 )
