+++ warehouse/src/main/kotlin/it/polito/wa2/group17/warehouse/dto/SellResponse.kt
@@ -5,6 +5,6 @@ import javax.validation.constraints.NotNull
 
 data class SellResponse(
     @field:NotNull val productID: Long,
-    @field:NotNull @field:Min(0) val quantity: Long,
+    @field:NotNull @field:Min(0) val quantity: Int,
     @field:NotNull val warehouseID: Long
 )
